setup CommentsModal with multiline text field. Next, setup a comments patch saga and router patch fn
